Refactor SPDX maker calls into separate file
