+++ sbom.py
@@ -1,0 +1,58 @@
+# SPDX-License-Identifier: Apache-2.0
+
+import hashlib
+import os
+
+from cmakefileapijson import parseReply
+from spdx.builder import BuilderConfig, makeSPDX
+
+def makeCmakeSpdx(replyIndexPath, srcRootDir, spdxOutputDir, spdxNamespacePrefix):
+    # get CMake info from build
+    cm = parseReply(replyIndexPath)
+
+    # create SPDX file for sources
+    srcSpdxPath = os.path.join(spdxOutputDir, "sources.spdx")
+    srcCfg = BuilderConfig()
+    srcCfg.documentName = "sources"
+    srcCfg.documentNamespace = os.path.join(spdxNamespacePrefix, "sources")
+    srcCfg.packageName = "sources"
+    srcCfg.spdxID = "SPDXRef-sources"
+    srcCfg.doSHA256 = True
+    srcCfg.scandir = srcRootDir
+    srcCfg.excludeDirs.append(cm.paths_build)
+    srcPkg = makeSPDX(srcCfg, srcSpdxPath)
+    if srcPkg:
+        print(f"Saved sources SPDX to {srcSpdxPath}")
+    else:
+        print(f"Couldn't generate sources SPDX file")
+
+    # get hash of sources SPDX file, to use for build doc's extRef
+    hSHA256 = hashlib.sha256()
+    with open(srcSpdxPath, 'rb') as f:
+        buf = f.read()
+        hSHA256.update(buf)
+    srcSHA256 = hSHA256.hexdigest()
+
+    # create SPDX file for build
+    buildSpdxPath = os.path.join(spdxOutputDir, "build.spdx")
+    buildCfg = BuilderConfig()
+    buildCfg.documentName = "build"
+    buildCfg.documentNamespace = os.path.join(spdxNamespacePrefix, "build")
+    buildCfg.packageName = "build"
+    buildCfg.spdxID = "SPDXRef-build"
+    buildCfg.doSHA256 = True
+    buildCfg.scandir = cm.paths_build
+
+    # add external document ref to sources SPDX file
+    buildCfg.extRefs = [("DocumentRef-sources", srcCfg.documentNamespace, "SHA256", srcSHA256)]
+
+    # exclude CMake file-based API responses -- presume only used for this
+    # SPDX generation scan, not for actual build artifact
+    buildExcludeDir = os.path.join(cm.paths_build, ".cmake", "api")
+    buildCfg.excludeDirs.append(buildExcludeDir)
+
+    buildPkg = makeSPDX(buildCfg, buildSpdxPath)
+    if buildPkg:
+        print(f"Saved build SPDX to {buildSpdxPath}")
+    else:
+        print(f"Couldn't generate build SPDX file")
